+++ R/jc_tombola.R
@@ -4,6 +4,7 @@
 #'
 #' @param data Data frame withe members and their information.
 #' @param members Columns with the members names.
+#' @param papers Number of paper by meeting 
 #' @param group Column for arrange the group.
 #' @param gr_lvl Levels in the groups for the arrange. See details.
 #' @param status Column with the status of the members.
@@ -25,6 +26,7 @@
 
 jc_tombola <- function(data
                        , members
+                       , papers = 1
                        , group
                        , gr_lvl
                        , status
@@ -33,41 +35,57 @@ jc_tombola <- function(data
                        , date
                        , seed = NULL
                        ){
-
+  
 grp <- NULL
 
 date <- as.Date(date)
-members <- as.name(members)
-group <- as.name(group)
-status <- as.name(status)
 
-gr_lvl <- structure(as.character(gr_lvl),
-                 names = as.character(1:length(gr_lvl)))
+if(is.null(seed)){ set.seed(date) } else {set.seed(seed)} 
 
-if(is.null(seed)){
-  set.seed(date)
-} else {set.seed(seed)} # replicate results
+dt <- data %>% 
+  mutate(across(c({{members}}, {{group}}, {{status}}), as.character)) 
 
-jc <- data %>%
-  dplyr::filter(!!status %in% st_lvl) %>%
-  dplyr::mutate(grp = dplyr::case_when(
-    !!group %in% gr_lvl ~ 1,
-    !(!!group %in% gr_lvl) ~ 2,
-  )) %>%
-  tidyr::uncount(grp) %>%
-  dplyr::group_by(!!members) %>%
-  dplyr::mutate(!!group := if(dplyr::n() > 1) {paste0(dplyr::row_number())}
-                else {paste0(!!group)}) %>%
-  dplyr::mutate(!!group := dplyr::recode(!!group,  !!!gr_lvl))
+gr.lvl <- dt %>% 
+  select({{group}}) %>% 
+  mutate( {{group}} := case_when(
+    {{group}} %in% gr_lvl ~ {{group}}
+    , TRUE ~ paste0(gr_lvl, collapse = ' ')
+  )) %>% 
+  tidyr::separate_rows({{group}}) %>% 
+  unique() %>% 
+  rownames_to_column() %>% 
+  select({{group}}, .data$rowname) %>% 
+  deframe()
+  
+jc <- dt %>%
+  dplyr::filter(.data[[status]] %in% st_lvl) %>%
+  mutate( {{group}} := case_when(
+    .data[[group]] %in% gr_lvl ~ .data[[group]]
+    , TRUE ~ paste0(gr_lvl, collapse = ' ')
+  )) %>% 
+  tidyr::separate_rows({{group}}, sep = ' ') %>% 
+  mutate(ngrp = case_when(
+    .data[[group]] %in% gr_lvl ~ gr.lvl[.data[[group]]]
+    ))
 
-tb <- jc %>%
-  dplyr::group_by(!!group) %>%
-  dplyr::mutate(grp := sample.int(dplyr::n())) %>%
-  dplyr::arrange(grp, !!group) %>%
-  dplyr::select(grp, !!group, !!members) %>%
-  dplyr::ungroup(!!group) %>%
+tb <- jc %>% 
+  {
+  
+    if(papers == length(gr_lvl)) {  
+      dplyr::group_by(.data = ., .data[[group]] ) } else {.}
+  
+  } %>% 
+  dplyr::mutate(grp = sample.int(dplyr::n())) %>% 
+  dplyr::arrange(.data$grp,  {{group}}) %>%
+  {
+
+    if(papers == length(gr_lvl)) {
+      dplyr::ungroup(x = ., .data[[group]] ) } else {.}
+
+  } %>%
+  dplyr::select(.data$grp, {{group}}, members) %>% 
   dplyr::mutate(grp = date + rep(seq(0, nrow(.)/length(gr_lvl)*frq, by = frq)
-                                 , each = length(gr_lvl)
+                                 , each = papers
                                  , len = nrow(.))) %>%
   dplyr::rename(Date = grp, Leader = members)